Start Skeleton enemy
just basic walking around
