+++ Assets/C#/EnemyScripts/SkeletonEnemy.cs
@@ -1,0 +1,42 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SkeletonEnemy : PatrolGroundEnemy {
+
+    private Transform target;
+
+
+
+    // Use this for initialization
+    void Start()
+    {
+        //use this for PatrolAround
+        base.__init__();
+
+    }
+
+    public override IEnumerator Attack()
+    {
+        return null;
+    }
+
+    public override void Movement()
+    {
+        if (!isPatrolling && patrolPositions.Length > 0)
+        {
+            StartCoroutine(PatrolAround());
+        }
+        
+    }
+
+    private void OnTriggerEnter(Collider other)
+    {
+
+    }
+
+
+
+
+}
