+++ Assets/C#/EnemyScripts/SkeletonEnemy.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3df902d5cfdcde14a83c03d4ba1a3f42
-timeCreated: 1490310513
+guid: 896f59fdb2a98f340ae60e05c25de558
+timeCreated: 1490741404
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
