+++ Assets/Materials/Skeleton.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a85cc1e83bd169e42ac4f1b726b14e1e
+timeCreated: 1490742446
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
