+++ Assets/Prefabs/Skeleton.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac9f3c0e7ea16a842a92dcefb742cf1d
+timeCreated: 1490742420
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
